+++ src/Console/Command/SystemCommands.php
@@ -130,7 +130,7 @@ class SystemCommands extends Commands
             ]
         ];
 
-        $link = $this->color->stringColor("https://solitalframework.com/docs/", "cyan", null);
+        $link = $this->color->stringColor("http://solitalframework.com/", "cyan", null);
         $thanks = "Thank you for using Solital, you can see the full documentation at $link\n";
 
         print_r($this->color->stringColor("\nSolital framework - Fast, easy and practical\n", "yellow", null, true));
@@ -293,7 +293,10 @@ class SystemCommands extends Commands
             "login",
             "remove-login",
             "forgot",
-            "remove-forgot"
+            "remove-forgot",
+            "minify-css",
+            "minify-js",
+            "minify-all"
         ];
 
         $desc = [
@@ -305,7 +308,10 @@ class SystemCommands extends Commands
             "Create classes for login",
             "Removes the components created for login",
             "Create classes for forgot password",
-            "Removes the components created for forgot password"
+            "Removes the components created for forgot password",
+            "Minimize CSS files",
+            "Minimize JavaScript files",
+            "Minimize CSS and JavaScript files"
         ];
 
         return [