+++ src/Exceptions/ContainerNotFoundException.php
@@ -2,13 +2,19 @@
 
 namespace Solital\Core\Exceptions;
 
+use ModernPHPException\ModernPHPException;
 use Psr\Container\NotFoundExceptionInterface;
 
 class ContainerNotFoundException extends \Exception implements NotFoundExceptionInterface
 {
+    /**
+     * @var ModernPHPException
+     */
+    private ModernPHPException $exception;
+
     // Redefine the exception so message isn't optional
-    public function __construct($id, $code = 0, \Exception $previous = null)
+    public function __construct($id)
     {
-        parent::__construct("Dependency \"" . $id . "\" not found. Make sure it is properly injected in this domain", $code, $previous);
+        $this->exception->errorHandler(404, "Dependency \"" . $id . "\" not found", __FILE__, __LINE__);
     }
 }