+++ CHANGELOG.md
@@ -1,6 +1,71 @@
 # Released Notes
 
-## v2.0.0-rc4 - (2021-04-XX)
+## v2.3.0 - (2021-07-xx)
+
+### Added
+
+
+### Change
+
+
+### Fixed
+
+--------------------------------------------------------------------------
+
+## v2.2.2 - (2021-06-29)
+
+### Added
+
+- Added test for Cache
+
+### Change
+
+- Change Bootstrap CDN
+
+### Fixed
+
+- Fixed bug Database in `Guardian` class
+- Fixed return type in `Cache` class
+--------------------------------------------------------------------------
+
+## v2.2.1 - (2021-06-28)
+
+### Fixed
+
+- Fixed bug in Wolf Template
+--------------------------------------------------------------------------
+
+## v2.2.0 - (2021-06-19)
+
+### Added
+
+- Added Added minify by Vinci
+- Added method to get data from `php://input`
+
+### Removed
+
+- Removed `WolfMinify` class
+- Removed unnecessary comments
+--------------------------------------------------------------------------
+
+## v2.1.0 - (2021-04-23)
+
+### Added
+
+- Added new helper: `console_log()`
+- Added support for Sodium encryption
+- Added `respect\validation` package
+- Added `isBase64` and `identical` methods in `Valid`
+- Added new tests
+
+### Fixed
+
+- Fixed namespace in helper output 
+- Fixed PSR-14
+- Fixed validation in `Valid` and `is_json`
+
+--------------------------------------------------------------------------
+## v2.0.0 - (2021-04-15)
 
 ### Added